NM modified files.

+++ satreps_nas_upload.py
@@ -486,9 +486,13 @@ for subdir in opts.subdir:
                     sys.stderr.flush()
         if len(os.listdir(srcdir)) == 0:
             if not srcdir.lower() in keep_folder_lower:
-                os.rmdir(srcdir)
-                if opts.debug and not os.path.exists(srcdir):
-                    sys.stderr.write('Removed {}\n'.format(srcdir))
+                if os.path.isdir(locdir):
+                    os.rmdir(srcdir)
+                    if opts.debug and not os.path.exists(srcdir):
+                        sys.stderr.write('Removed {}\n'.format(srcdir))
+                        sys.stderr.flush()
+                else:
+                    sys.stderr.write('Warning, no such directory >>> {}\n'.format(locdir))
                     sys.stderr.flush()
 # Remove empty directories
 for subdir in opts.subdir:
@@ -496,11 +500,17 @@ for subdir in opts.subdir:
         curdir = os.path.relpath(root,opts.srcdir)
         if curdir == os.curdir:
             srcdir = opts.srcdir
+            locdir = opts.locdir
         else:
             srcdir = os.path.join(opts.srcdir,curdir)
+            locdir = os.path.join(opts.locdir,curdir)
         if len(os.listdir(srcdir)) == 0:
             if not srcdir.lower() in keep_folder_lower:
-                os.rmdir(srcdir)
-                if opts.debug and not os.path.exists(srcdir):
-                    sys.stderr.write('Removed {}\n'.format(srcdir))
+                if os.path.isdir(locdir):
+                    os.rmdir(srcdir)
+                    if opts.debug and not os.path.exists(srcdir):
+                        sys.stderr.write('Removed {}\n'.format(srcdir))
+                        sys.stderr.flush()
+                else:
+                    sys.stderr.write('Warning, no such directory >>> {}\n'.format(locdir))
                     sys.stderr.flush()
